requirements.txt
streamlit
opencv-python
torch
facenet-pytorch
numpy
pillow
pandas